refactor: reorganize types; rename type modules
