+++ packages/core/src/engine/PropagateUpdate.ts
@@ -4,14 +4,11 @@ import {
   exprToNumber,
 } from "engine/EngineUtils";
 import { getShapeName } from "renderer/ShapeDef";
-import {
-  Path,
-  IPropertyPath,
-  IAccessPath,
-  Translation,
-} from "types/shapeEvalTypes";
-import { Shape, Value } from "types/shapeTypes";
-import { State, LabelCache } from "types/stateTypes";
+import { Translation } from "types/value";
+import { Shape } from "types/shape";
+import { Value } from "types/value";
+import { State, LabelCache } from "types/state";
+import { Path, IPropertyPath, IAccessPath } from "types/style";
 import { retrieveLabel } from "utils/CollectLabels";
 
 /**
