+++ packages/core/src/parser/Domain.ne
@@ -7,9 +7,8 @@
 import * as moo from "moo";
 import { concat, compact, flatten, last } from 'lodash'
 import { optional, tokensIn, basicSymbols, rangeOf, rangeBetween, rangeFrom, nth, convertTokenId } from 'parser/ParserUtil'
-import { ASTNode } from "types/ASTTypes";
-import { DomainProg, TypeDecl, PredicateDecl, FunctionDecl, ConstructorDecl, PreludeDecl, NotationDecl, SubTypeDecl, TypeConstructor, Type, Arg, Prop } from "types/domainASTTypes";
-import { IStringLit } from "types/shapeEvalTypes";
+import { ASTNode, IStringLit } from "types/ast";
+import { DomainProg, TypeDecl, PredicateDecl, FunctionDecl, ConstructorDecl, PreludeDecl, NotationDecl, SubTypeDecl, TypeConstructor, Type, Arg, Prop } from "types/domain";
 
 // NOTE: ordering matters here. Top patterns get matched __first__
 const lexer = moo.compile({
