+++ packages/core/src/renderer/dragUtils.ts
@@ -1,6 +1,6 @@
 import { updateVaryingValues } from "engine/PropagateUpdate";
-import { Properties, Shape } from "types/shapeTypes";
-import { State } from "types/stateTypes";
+import { Properties, Shape } from "types/shape";
+import { State } from "types/state";
 
 /**
  * Retrieve data from drag events and update varying state accordingly
