--- packages/core/src/types/shapeTypes.ts
@@ -1,78 +1,0 @@
-import {
-  IFloatV,
-  IIntV,
-  IBoolV,
-  IStrV,
-  IPtV,
-  IPathDataV,
-  IPtListV,
-  IColorV,
-  IPaletteV,
-  IFileV,
-  IStyleV,
-  IListV,
-  IVectorV,
-  IMatrixV,
-  ITupV,
-  ILListV,
-  IHMatrixV,
-  IPolygonV,
-  IPt,
-  ICubicBez,
-  ICubicBezJoin,
-  IQuadBez,
-  IQuadBezJoin,
-} from "./shapeEvalTypes";
-
-export type Shape = IShape;
-
-export type Properties = { [k: string]: Value<number> };
-
-/**
- * A shape (Graphical Primitive Instance, aka GPI) in penrose has a type (_e.g._ `Circle`) and a set of properties (_e.g._ `center`). This type is specifically used for rendering. See {@link GPI} for the version used for the runtime.
- */
-interface IShape {
-  shapeType: string;
-  properties: Properties;
-}
-
-/**
- * A value in the penrose system.
- */
-export type Value<T> =
-  | IFloatV<T>
-  | IIntV
-  | IBoolV<T>
-  | IStrV<T>
-  | IPtV<T>
-  | IPathDataV<T>
-  | IPtListV<T>
-  | IColorV<T>
-  | IPaletteV<T>
-  | IFileV<T>
-  | IStyleV<T>
-  | IListV<T>
-  | IVectorV<T>
-  | IMatrixV<T>
-  | ITupV<T>
-  | ILListV<T>
-  | IHMatrixV<T>
-  | IPolygonV<T>;
-
-export type Elem<T> =
-  | IPt<T>
-  | ICubicBez<T>
-  | ICubicBezJoin<T>
-  | IQuadBez<T>
-  | IQuadBezJoin<T>;
-
-export type SubPath<T> = IClosed<T> | IOpen<T>;
-interface IClosed<T> {
-  tag: "Closed";
-  contents: Elem<T>[];
-}
-
-interface IOpen<T> {
-  tag: "Open";
-  contents: Elem<T>[];
-}
