+++ packages/core/src/types/value.ts
@@ -1,7 +1,6 @@
-import { VarAD } from "./adTypes";
-import { ASTNode, Identifier } from "./ASTTypes";
+import { VarAD } from "./ad";
 import { StyleError } from "./errors";
-import { SubPath, Value } from "./shapeTypes";
+import { Expr } from "./style";
 
 /**
  * The input parameters to computations/objectives/constraints in Style. It can be either an entire shape (`IGPI`) or a value (`IVal`).
@@ -82,226 +81,51 @@ export interface IPending<T> {
   contents: Value<T>;
 }
 
-export type Expr =
-  // | IIntLit
-  | AnnoFloat
-  | IStringLit
-  | IBoolLit
-  | Path // NOTE: changed from EPath
-  | ICompApp
-  | IObjFn
-  | IConstrFn
-  // | IAvoidFn // TODO: unimplemented
-  | IBinOp
-  | IUOp
-  | IList
-  | ITuple
-  | IVector
-  | IMatrix
-  | IVectorAccess
-  | IMatrixAccess
-  | IListAccess
-  | GPIDecl
-  | ILayering
-  | IPluginAccess;
-// | IThenOp; // TODO: deprecated transformation exprs
+/**
+ * A value in the penrose system.
+ */
+export type Value<T> =
+  | IFloatV<T>
+  | IIntV
+  | IBoolV<T>
+  | IStrV<T>
+  | IPtV<T>
+  | IPathDataV<T>
+  | IPtListV<T>
+  | IColorV<T>
+  | IPaletteV<T>
+  | IFileV<T>
+  | IStyleV<T>
+  | IListV<T>
+  | IVectorV<T>
+  | IMatrixV<T>
+  | ITupV<T>
+  | ILListV<T>
+  | IHMatrixV<T>
+  | IPolygonV<T>;
 
-export interface IIntLit {
-  tag: "IntLit";
-  contents: number;
+export type Elem<T> =
+  | IPt<T>
+  | ICubicBez<T>
+  | ICubicBezJoin<T>
+  | IQuadBez<T>
+  | IQuadBezJoin<T>;
+
+export type SubPath<T> = IClosed<T> | IOpen<T>;
+interface IClosed<T> {
+  tag: "Closed";
+  contents: Elem<T>[];
 }
 
-// interface IAFloat {
-//     tag: "AFloat";
-//     contents: AnnoFloat;
-// }
-
-export interface IStringLit extends ASTNode {
-  tag: "StringLit";
-  contents: string;
-}
-
-export interface IBoolLit extends ASTNode {
-  tag: "BoolLit";
-  contents: boolean;
-}
-
-export interface ICompApp extends ASTNode {
-  tag: "CompApp";
-  name: Identifier;
-  args: Expr[];
-}
-
-export interface IObjFn extends ASTNode {
-  tag: "ObjFn";
-  name: Identifier;
-  args: Expr[];
-}
-
-export interface IConstrFn extends ASTNode {
-  tag: "ConstrFn";
-  name: Identifier;
-  args: Expr[];
-}
-
-export interface IAvoidFn {
-  tag: "AvoidFn";
-  contents: [string, Expr[]];
-}
-
-export type BinaryOp = "BPlus" | "BMinus" | "Multiply" | "Divide" | "Exp";
-
-// NOTE: unary + operator not parsed, as they don't change values
-export type UnaryOp = "UMinus";
-export interface IBinOp extends ASTNode {
-  tag: "BinOp";
-  op: BinaryOp;
-  left: Expr;
-  right: Expr;
-}
-export interface IUOp extends ASTNode {
-  tag: "UOp";
-  op: UnaryOp;
-  arg: Expr;
-}
-
-export interface IList extends ASTNode {
-  tag: "List";
-  contents: Expr[];
-}
-
-export interface ITuple extends ASTNode {
-  tag: "Tuple";
-  contents: [Expr, Expr];
-}
-
-export interface IVector extends ASTNode {
-  tag: "Vector";
-  contents: Expr[];
-}
-
-export interface IMatrix extends ASTNode {
-  tag: "Matrix";
-  contents: Expr[];
-}
-
-export interface IVectorAccess extends ASTNode {
-  tag: "VectorAccess";
-  contents: [Path, Expr];
-}
-
-export interface IMatrixAccess extends ASTNode {
-  tag: "MatrixAccess";
-  contents: [Path, Expr[]];
-}
-
-export interface IListAccess extends ASTNode {
-  tag: "ListAccess";
-  contents: [Path, number];
-}
-
-export interface GPIDecl extends ASTNode {
-  tag: "GPIDecl";
-  shapeName: Identifier;
-  properties: PropertyDecl[];
-}
-
-export interface ILayering extends ASTNode {
-  tag: "Layering";
-  below: Path;
-  above: Path;
-}
-
-export interface IPluginAccess extends ASTNode {
-  tag: "PluginAccess";
-  contents: [string, Expr, Expr];
-}
-
-export interface IThenOp {
-  tag: "ThenOp";
-  contents: [Expr, Expr];
-}
-
-export type AnnoFloat = IFix | IVary | IVaryAD;
-
-export interface IFix extends ASTNode {
-  tag: "Fix";
-  contents: number;
-}
-
-export interface IVary extends ASTNode {
-  tag: "Vary";
-}
-
-export interface IVaryAD extends ASTNode {
-  tag: "VaryAD";
-  contents: VarAD;
-}
-
-export interface PropertyDecl extends ASTNode {
-  tag: "PropertyDecl";
-  name: Identifier;
-  value: Expr;
-}
-
-// TODO: check how the evaluator/compiler should interact with ASTNode
-export type Path =
-  | IFieldPath
-  | IPropertyPath
-  | IAccessPath
-  | LocalVar
-  | IInternalLocalVar;
-// LocalVar is only used internally by the compiler
-// Unused
-// | ITypePropertyPath;
-
-export interface IFieldPath extends ASTNode {
-  tag: "FieldPath";
-  name: BindingForm;
-  field: Identifier;
-}
-
-export interface IPropertyPath extends ASTNode {
-  tag: "PropertyPath";
-  name: BindingForm;
-  field: Identifier;
-  property: Identifier;
-}
-
-export interface IAccessPath extends ASTNode {
-  tag: "AccessPath";
-  path: Path;
-  indices: Expr[];
-}
-
-// COMBAK: This is named inconsistently since the parser calls it `LocalVar`, should be ILocalVar
-export interface LocalVar extends ASTNode {
-  tag: "LocalVar";
-  contents: Identifier;
-}
-
-export interface IInternalLocalVar extends ASTNode {
-  // Note, better to not extend ASTNode as it's only used internally by compiler, but breaks parser otherwise
-  tag: "InternalLocalVar";
-  contents: string;
+interface IOpen<T> {
+  tag: "Open";
+  contents: Elem<T>[];
 }
 
 // Unused
 // interface ITypePropertyPath {
 //   tag: "TypePropertyPath";
 // }
-
-export type BindingForm = SubVar | StyVar;
-
-export interface SubVar extends ASTNode {
-  tag: "SubVar";
-  contents: Identifier;
-}
-
-export interface StyVar extends ASTNode {
-  tag: "StyVar";
-  contents: Identifier;
-}
 
 /** A floating point number **/
 export interface IFloatV<T> {
